Korean translation update

+++ src/main/java/net/dries007/tfc/common/blocks/plant/fruit/DeadCaneBlock.java
@@ -22,7 +22,7 @@ import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.phys.BlockHitResult;
 
 import net.dries007.tfc.common.TFCTags;
-import net.dries007.tfc.common.blockentities.TFCBlockEntities;
+import net.dries007.tfc.common.blockentities.TickCounterBlockEntity;
 import net.dries007.tfc.common.blocks.EntityBlockExtension;
 import net.dries007.tfc.common.blocks.ExtendedProperties;
 import net.dries007.tfc.util.calendar.ICalendar;
@@ -40,12 +40,13 @@ public class DeadCaneBlock extends SpreadingCaneBlock implements EntityBlockExte
     {
         if (random.nextInt(15) == 0 && !level.getBlockState(pos.above()).is(TFCTags.Blocks.ANY_SPREADING_BUSH))
         {
-            level.getBlockEntity(pos, TFCBlockEntities.TICK_COUNTER.get()).ifPresent(counter -> {
+            if (level.getBlockEntity(pos) instanceof TickCounterBlockEntity counter)
+            {
                 if (counter.getTicksSinceUpdate() > ICalendar.TICKS_IN_DAY * 80)
                 {
                     level.destroyBlock(pos, true);
                 }
-            });
+            }
         }
     }
